third_party/blink/web_tests/external/wpt/css/css-color/parsing/color-computed-relative-color-expected.txt

This is a testharness.js-based test.
Found 8 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Property color value 'rgb(from rebeccapurple b alpha r / g)'
  Colors do not match.\nActual:   color(srgb 0.6 0.00392157 0.4)\nExpected: color(srgb 0.6 1 0.4 / 0.2).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. lengths differ, expected 4 got 3
[FAIL] Property color value 'rgb(from rebeccapurple r r r / r)'
  Colors do not match.\nActual:   color(srgb 0.4 0.4 0.4)\nExpected: color(srgb 0.4 0.4 0.4 / 0.4).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. lengths differ, expected 4 got 3
[FAIL] Property color value 'rgb(from rebeccapurple alpha alpha alpha / alpha)'
  Colors do not match.\nActual:   color(srgb 0.00392157 0.00392157 0.00392157)\nExpected: color(srgb 1 1 1).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 1 +/- 0.01, expected 1 but got 0.00392157
[FAIL] Property color value 'rgb(from rgb(20%, 40%, 60%, 80%) b alpha r / g)'
  Colors do not match.\nActual:   color(srgb 0.6 0.00313725 0.2)\nExpected: color(srgb 0.6 0.8 0.2 / 0.4).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. lengths differ, expected 4 got 3
[FAIL] Property color value 'rgb(from rgb(20%, 40%, 60%, 80%) r r r / r)'
  Colors do not match.\nActual:   color(srgb 0.2 0.2 0.2)\nExpected: color(srgb 0.2 0.2 0.2 / 0.2).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. lengths differ, expected 4 got 3
[FAIL] Property color value 'rgb(from rgb(20%, 40%, 60%, 80%) alpha alpha alpha / alpha)'
  Colors do not match.\nActual:   color(srgb 0.00313725 0.00313725 0.00313725 / 0.8)\nExpected: color(srgb 0.8 0.8 0.8 / 0.8).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.8 +/- 0.01, expected 0.8 but got 0.00313725
[FAIL] Property color value 'hsl(from hsl(20 30 40 / 0.8) calc(h + 1) calc(s + 1) calc(l + 1) / calc(alpha + 0.01))'
  Colors do not match.\nActual:   color(srgb 0.88 1.556 1.92 / 0.81)\nExpected: color(srgb 0.54 0.37 0.28 / 0.81).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.54 +/- 0.01, expected 0.54 but got 0.88
[FAIL] Property color value 'hwb(from hwb(20 30 40 / 0.8) calc(h + 1) calc(w + 1) calc(b + 1) / calc(alpha + 0.01))'
  Colors do not match.\nActual:   color(srgb 0.481481 0.481481 0.481481 / 0.81)\nExpected: color(srgb 0.59 0.41 0.31 / 0.81).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 0.59 +/- 0.01, expected 0.59 but got 0.481481
Harness: the test ran to completion.
